@@ -13,47 +13,33 @@
             "email": "hugo@alliau.me"
         }
     ],
-    "scripts": {
-        "build-scoped-dependencies": [
-            "rm -fr scoped-dependencies/vendor",
-            "@composer install --working-dir scoped-dependencies --prefer-dist --no-dev --ansi",
-            "@composer dump-autoload --working-dir scoped-dependencies --classmap-authoritative --no-dev --ansi",
-            "@php -dmemory_limit=-1 ./scoped-dependencies/build/php-scoper.phar add-prefix --working-dir ./scoped-dependencies/build --output-dir .scoped --force --ansi"
-        ]
-    },
     "require": {
-        "php": ">=8.0",
+        "php": ">=8.1",
         "ext-zip": "*",
-        "psr/cache": "^2.0 || ^3.0",
-        "psr/log": "^1.0 || ^2.0 || ^3.0",
-        "symfony/framework-bundle": "^5.4 || ^6.0",
-        "symfony/form": "^5.4 || ^6.0",
-        "symfony/finder": "^5.4 || ^6.0",
-        "symfony/console": "^5.4 || ^6.0",
-        "symfony/mime": "^5.4 || ^6.0",
+        "league/flysystem": "^3.0",
+        "league/flysystem-aws-s3-v3": "^3.0.1",
+        "symfony/framework-bundle": "^6.0",
+        "symfony/form": "^6.0",
+        "symfony/finder": "^6.0",
+        "symfony/console": "^6.0",
+        "symfony/mime": "^6.0",
         "pimple/pimple": "^3.0",
-        "monolog/monolog": "~1.4|~2.0"
+        "monolog/monolog": "^2.0"
     },
     "require-dev": {
         "symplify/easy-coding-standard": "^10.1",
-        "symfony/yaml": "^5.4 || ^6.0",
+        "symfony/yaml": "^6.0",
         "rector/rector": "^0.14.2",
         "phpstan/phpstan": "^1.8",
         "phpstan/phpstan-phpunit": "^1.0",
         "phpunit/phpunit": "^9.5",
-        "aws/aws-sdk-php": "^3.261"
-    },
-    "conflict": {
-        "guzzlehttp/guzzle": "<6.0"
+        "aws/aws-sdk-php": "^3.269.8"
     },
     "autoload": {
         "psr-4": {
             "CKSource\\Bundle\\CKFinderBundle\\": "src",
             "CKSource\\CKFinder\\": "src/_connector"
-        },
-        "files": [
-            "scoped-dependencies/build/.scoped/vendor/scoper-autoload.php"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
