@@ -10,8 +10,9 @@ return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->paths([
         __DIR__ . '/src',
         __DIR__ . '/tests',
-        __DIR__ . '/playground/symfony-5.4/src',
         __DIR__ . '/playground/symfony-6.0/src',
+        __DIR__ . '/playground/symfony-6.1/src',
+        __DIR__ . '/playground/symfony-6.2/src',
     ]);
 
     $rectorConfig->skip([
